+++ CHANGES.rst
@@ -3,9 +3,21 @@
 Changelog
 =========
 
-vo.0.1 (10-11-2016)
+v0.0.2 (10-14-2016)
+-------------------
+
+- Second release.
+ - added support for model creation
+ - added database orm suppot with ormlite
+ - changed directory structure a little moving db.sqlite out of database folder
+ - change JSON file for model creation structure
+ - changed command create table to create model
+.. _Othree: https://github.com/othreecodes
+
+v0.0.1 (10-11-2016)
 -------------------
 
 - First release.
 
 .. _Othree: https://github.com/othreecodes
+